+++ loopx/cli_commands/quota.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import argparse
 from collections.abc import Callable, Mapping
-from dataclasses import dataclass
 from pathlib import Path
 
 from ..capabilities.explore.composition_frontier import (
@@ -43,17 +42,9 @@ from ..control_plane.quota.settlement_cli import (
     render_existing_heartbeat_receipt_payload,
 )
 from ..control_plane.quota.turn_envelope import build_turn_envelope
-from ..control_plane.runtime.status_projection_cache import (
-    load_status_projection_cache,
-    resolve_status_projection_cache_runtime_root,
-    write_status_projection_cache,
-)
 from ..control_plane.scheduler.execution_context import (
     GUIDED_START_TURN_RUNTIME_PROFILES,
     SchedulerExecutionContextResolution,
-    SchedulerRuntimeProfile,
-    scheduler_execution_context_for_runtime_profile,
-    scheduler_runtime_profile_for_execution_context,
 )
 from ..control_plane.todos.contract import normalize_todo_id
 from ..file_lock import lock_timeout_error_fields
@@ -78,20 +69,13 @@ from ..quota import (
     spend_quota_slot,
     void_quota_slot,
 )
-from ..status import AUTONOMOUS_REPLAN_PERIODIC_LOOKBACK, collect_status
-from ..turn_identity import mint_turn_instance_id, normalize_turn_instance_id
+from ..status import collect_status
 from ..upgrade import resolve_codex_app_automation_rrule
-from .lark_inbox import build_lark_operator_inbox_urgency_projector
+from .quota_context import QuotaCommandContext, prepare_quota_command_context
 from .quota_host_poll import attach_host_poll_receipt
 from .quota_monitor_poll import record_quota_monitor_poll_for_cli
 from .quota_registration import (
     register_quota_command as register_quota_command,  # noqa: PLC0414
-)
-from .quota_request import (
-    QUOTA_MONITOR_POLL_DETAIL_SECTIONS,
-    QUOTA_SHOULD_RUN_DETAIL_SECTIONS,
-    quota_detail_sections_from_args,
-    validate_quota_command_request,
 )
 
 PrintPayload = Callable[
@@ -108,229 +92,12 @@ QUOTA_EVENT_KINDS = {
     "spend-slot": "quota_spend",
     "void-slot": "quota_void",
 }
-QUOTA_SCHEDULER_COMMANDS = frozenset(
-    {
-        "should-run",
-        "monitor-poll",
-        "scheduler-ack",
-        "scheduler-ack-current",
-        "scheduler-fail-current",
-        "spend-slot",
-    }
-)
-
-
-@dataclass(frozen=True, slots=True)
-class _QuotaCommandContext:
-    runtime_root: Path
-    scan_roots: list[Path]
-    status_limit: int
-    status_goal_id: str | None
-    status_payload: dict[str, object]
-    cache_metadata: dict[str, object] | None
-    scheduler_context: Mapping[str, object] | SchedulerExecutionContextResolution | None
-    operator_inbox_urgency_projector: Callable[..., dict[str, object]]
-    detail_sections: frozenset[str]
-    heartbeat_turn_id: str | None
-
-
 def _heartbeat_receipt_settlement_bindings(
     event: Mapping[str, object],
 ) -> tuple[str | None, str | None]:
     return (
         heartbeat_receipt_settlement_todo_id(event),
         heartbeat_receipt_settlement_replan_obligation_id(event),
-    )
-
-
-def _scheduler_execution_context_from_args(
-    args: argparse.Namespace,
-) -> Mapping[str, object] | SchedulerExecutionContextResolution | None:
-    explicit_scheduler_fields = (
-        args.host_surface,
-        args.scheduler_owner,
-        args.execution_mode,
-    )
-    if args.codex_app and (args.runtime_profile or any(explicit_scheduler_fields)):
-        raise QuotaCommandValidationError(
-            "--codex-app cannot be combined with --runtime-profile, "
-            "--host-surface, --scheduler-owner, or --execution-mode"
-        )
-    if args.runtime_profile and any(explicit_scheduler_fields):
-        raise QuotaCommandValidationError(
-            "--runtime-profile cannot be combined with --host-surface, "
-            "--scheduler-owner, or --execution-mode"
-        )
-    runtime_profile = (
-        SchedulerRuntimeProfile.CODEX_APP_HEARTBEAT.value
-        if args.codex_app
-        else args.runtime_profile
-    )
-    if runtime_profile:
-        return scheduler_execution_context_for_runtime_profile(runtime_profile)
-    if any(explicit_scheduler_fields):
-        return {
-            "host_surface": args.host_surface,
-            "scheduler_owner": args.scheduler_owner,
-            "execution_mode": args.execution_mode,
-            "source": "quota_cli_invocation",
-        }
-    return None
-
-
-def _prepare_quota_command_context(
-    args: argparse.Namespace,
-    *,
-    registry_path: Path,
-    runtime_root_arg: str | None,
-) -> _QuotaCommandContext:
-    command = args.quota_command
-    if bool(getattr(args, "turn_envelope", False)) and command != "should-run":
-        raise QuotaCommandValidationError(
-            "--turn-envelope is only valid with `quota should-run`"
-        )
-    requested_details = set(getattr(args, "include_details", None) or ())
-    if requested_details and command not in {"should-run", "monitor-poll"}:
-        raise QuotaCommandValidationError(
-            "--include-detail is only valid with `quota should-run` or "
-            "`quota monitor-poll`"
-        )
-    if requested_details and "all" not in requested_details:
-        allowed_details = set(
-            QUOTA_MONITOR_POLL_DETAIL_SECTIONS
-            if command == "monitor-poll"
-            else QUOTA_SHOULD_RUN_DETAIL_SECTIONS
-        )
-        unsupported_details = sorted(requested_details - allowed_details)
-        if unsupported_details:
-            raise QuotaCommandValidationError(
-                f"`quota {command}` does not accept --include-detail "
-                f"{', '.join(unsupported_details)}"
-            )
-    if (
-        bool(getattr(args, "include_scheduler_detail", False))
-        and command != "should-run"
-    ):
-        raise QuotaCommandValidationError(
-            "--include-scheduler-detail is only valid with `quota should-run`"
-        )
-    if bool(getattr(args, "record_host_poll", False)) and command != "should-run":
-        raise QuotaCommandValidationError(
-            "--record-host-poll is only valid with `quota should-run`"
-        )
-
-    begin_turn = bool(getattr(args, "begin_turn", False))
-    try:
-        heartbeat_turn_id = normalize_turn_instance_id(
-            getattr(args, "turn_instance_id", None)
-        )
-    except ValueError as exc:
-        raise QuotaCommandValidationError(str(exc)) from exc
-    if heartbeat_turn_id and command not in {
-        "should-run",
-        "monitor-poll",
-        "spend-slot",
-    }:
-        raise QuotaCommandValidationError(
-            "--turn-instance-id is only valid with `quota should-run`, "
-            "`quota monitor-poll`, or `quota spend-slot`"
-        )
-    if getattr(args, "replan_obligation_id", None) and command != "spend-slot":
-        raise QuotaCommandValidationError(
-            "--replan-obligation-id is only valid with `quota spend-slot`"
-        )
-    if getattr(args, "replan_obligation_id", None) and getattr(args, "todo_id", None):
-        raise QuotaCommandValidationError(
-            "--replan-obligation-id cannot be combined with --todo-id"
-        )
-    if heartbeat_turn_id and not args.agent_id:
-        raise QuotaCommandValidationError(
-            "turn-scoped quota settlement requires --agent-id"
-        )
-    scheduler_context = (
-        _scheduler_execution_context_from_args(args)
-        if command in QUOTA_SCHEDULER_COMMANDS
-        else None
-    )
-    validate_quota_command_request(args)
-    if begin_turn:
-        profile = scheduler_runtime_profile_for_execution_context(scheduler_context)
-        if profile not in GUIDED_START_TURN_RUNTIME_PROFILES:
-            raise QuotaCommandValidationError(
-                "--begin-turn requires runtime-profile codex_app_heartbeat "
-                "or codex_app_ssh_goal"
-            )
-        heartbeat_turn_id = mint_turn_instance_id(prefix="guided-start")
-    if heartbeat_turn_id and command == "should-run" and bool(args.dry_run):
-        raise QuotaCommandValidationError(
-            "turn-scoped `quota should-run` cannot use --dry-run"
-        )
-
-    scan_roots = [Path(item).expanduser() for item in args.scan_path]
-    if not scan_roots:
-        scan_roots = [Path(args.scan_root).expanduser()]
-    status_limit = max(0, args.limit)
-    if command in QUOTA_SCHEDULER_COMMANDS:
-        status_limit = max(status_limit, AUTONOMOUS_REPLAN_PERIODIC_LOOKBACK)
-    runtime_root = resolve_status_projection_cache_runtime_root(
-        registry_path=registry_path,
-        runtime_root_override=runtime_root_arg,
-    )
-    operator_inbox_urgency_projector = build_lark_operator_inbox_urgency_projector(
-        runtime_root_arg=runtime_root,
-    )
-    status_goal_id = args.goal_id if command not in {"status", "plan"} else None
-    projection_cache_ttl_seconds = int(
-        getattr(args, "projection_cache_ttl_seconds", 120)
-    )
-    status_payload = None
-    cache_metadata = None
-    if bool(getattr(args, "use_projection_cache", False)):
-        status_payload, cache_metadata = load_status_projection_cache(
-            registry_path=registry_path,
-            runtime_root=runtime_root,
-            scan_roots=scan_roots,
-            limit=status_limit,
-            include_task_graph=False,
-            goal_id=status_goal_id,
-            max_age_seconds=projection_cache_ttl_seconds,
-            available_capabilities=args.available_capabilities,
-        )
-    if status_payload is None:
-        status_payload = collect_status(
-            registry_path=registry_path,
-            runtime_root_override=runtime_root_arg,
-            scan_roots=scan_roots,
-            limit=status_limit,
-            goal_id=status_goal_id,
-            available_capabilities=args.available_capabilities,
-        )
-        if bool(getattr(args, "write_projection_cache", False)):
-            cache_metadata = write_status_projection_cache(
-                registry_path=registry_path,
-                runtime_root=runtime_root,
-                scan_roots=scan_roots,
-                limit=status_limit,
-                include_task_graph=False,
-                goal_id=status_goal_id,
-                payload=status_payload,
-                max_age_seconds=projection_cache_ttl_seconds,
-                available_capabilities=args.available_capabilities,
-            )
-    elif isinstance(status_payload.get("projection_cache"), dict):
-        cache_metadata = dict(status_payload["projection_cache"])
-
-    return _QuotaCommandContext(
-        runtime_root=runtime_root,
-        scan_roots=scan_roots,
-        status_limit=status_limit,
-        status_goal_id=status_goal_id,
-        status_payload=status_payload,
-        cache_metadata=cache_metadata,
-        scheduler_context=scheduler_context,
-        operator_inbox_urgency_projector=operator_inbox_urgency_projector,
-        detail_sections=quota_detail_sections_from_args(args),
-        heartbeat_turn_id=heartbeat_turn_id,
     )
 
 
@@ -643,9 +410,9 @@ def handle_quota_command(
     heartbeat_receipt_ready = False
     heartbeat_stall_observation = "not_evaluated"
     detail_sections: frozenset[str] = frozenset()
-    context: _QuotaCommandContext | None = None
+    context: QuotaCommandContext | None = None
     try:
-        context = _prepare_quota_command_context(
+        context = prepare_quota_command_context(
             args,
             registry_path=registry_path,
             runtime_root_arg=runtime_root_arg,
